Edit achievement : date format fix and refactor
