+++ app/src/main/java/org/ole/planet/myplanet/ui/userprofile/UserProfileFragment.java
@@ -68,7 +68,7 @@ public class UserProfileFragment extends Fragment {
         RealmUserModel model = mRealm.copyToRealmOrUpdate(handler.getUserModel());
         ((TextView) v.findViewById(R.id.txt_name)).setText(String.format("%s %s %s", model.getFirstName(), model.getMiddleName(), model.getLastName()));
         ((TextView) v.findViewById(R.id.txt_email)).setText(Utilities.checkNA(model.getEmail()));
-        String dob = TextUtils.isEmpty(model.getDob())? "N/A" : TimeUtils.getformatedDate(model.getDob(), "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'");
+        String dob = TextUtils.isEmpty(model.getDob())? "N/A" : TimeUtils.getFormatedDate(model.getDob(), "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'");
         ((TextView) v.findViewById(R.id.txt_dob)).setText(dob);
         Utilities.loadImage(model.getUserImage(), (ImageView) v.findViewById(R.id.image));
         final LinkedHashMap<String, String> map = new LinkedHashMap<String, String>();
