+++ app/src/main/java/org/ole/planet/myplanet/utilities/Utilities.java
@@ -61,20 +61,6 @@ public class Utilities {
         return getUrl() + "/_users/" + userId + "/" + imageName;
     }
 
-    public static String currentDate() {
-        Calendar c = Calendar.getInstance();
-        SimpleDateFormat dateformat = new SimpleDateFormat("EEE dd, MMMM yyyy");
-        String datetime = dateformat.format(c.getTime());
-        return datetime;
-    }
-
-    public static String formatDate(long date) {
-        Calendar c = Calendar.getInstance();
-        SimpleDateFormat dateformat = new SimpleDateFormat("EEE dd, MMMM yyyy");
-        String datetime = dateformat.format(date);
-        return datetime;
-    }
-
 
     public static void openDownloadService(Context context, ArrayList urls) {
         Intent intent = new Intent(context, MyDownloadService.class);
